requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
unstructured
python-docx
chromadb
pypdf
python-dotenv

